Add files via upload
Updated the SDFITS HDU construction to remove the second binary fits table place holder named NOISE. Not needed.
Also, modified the construction of the fits file name so that if this is called with '.' or './' as the directory path, it will
extract the basename of the absolute path to use as the root for the fits file name. E.g. calling with ./scan6 and 
./ when in scan6 will both result in scan6.fits as the output fits file.